services/adk_agent/pyproject.toml
[project]
name = "ssyk-adk-agent"
version = "0.1.0"
description = "ADK Agent for SSYK"
requires-python = ">=3.12"
dependencies = [
    "google-adk",
    "mcp>=1.0.0",
    "httpx",
    "matplotlib>=3.8.0",
    "seaborn>=0.13.0",
    "pandas>=2.0.0"
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/ssyk_advisor"]
